+++ sdk/core/test/integration/NetworkMember.test.ts
@@ -4,7 +4,6 @@ import { expect } from 'chai'
 import request from 'supertest'
 import { decode as jwtDecode } from 'jsonwebtoken'
 import { Affinity } from '@affinidi/common'
-import { DidAuthService } from '@affinidi/affinidi-did-auth-lib'
 import { buildVCV1Unsigned, buildVCV1Skeleton } from '@affinidi/vc-common'
 import { VCSPhonePersonV1, getVCPhonePersonV1Context } from '@affinidi/vc-data'
 import UserManagementService from '../../src/services/UserManagementService'
@@ -363,21 +362,6 @@ describe('CommonNetworkMember', () => {
     expect(didRequestTokenResponse.status).to.equal(200)
     expect(didRequestTokenResponse.body).to.exist
 
-    const didRequestToken = didRequestTokenResponse.body
-
-    const holderEncryptedSeed = ISSUER_ELEM_SEED
-    const holderPassword = password
-    const holderDidAuthServiceOptions = {
-      encryptedSeed: holderEncryptedSeed,
-      encryptionKey: holderPassword,
-    }
-    const didAuthService = new DidAuthService(holderDidAuthServiceOptions)
-    const didResponseToken = await didAuthService.createDidAuthResponseToken(didRequestToken)
-
-    expect(didResponseToken).to.exist
-
-    const accessToken = didResponseToken
-
     const credId = new Date().toISOString()
     const unsignedCredential = buildVCV1Unsigned({
       skeleton: buildVCV1Skeleton<VCSPhonePersonV1>({
@@ -396,10 +380,7 @@ describe('CommonNetworkMember', () => {
       expirationDate: new Date(new Date().getTime() + 10 * 60 * 1000).toISOString(),
     })
 
-    const revokableUnsignedCredential = await commonNetworkMember.buildRevocationListStatus(
-      unsignedCredential,
-      accessToken,
-    )
+    const revokableUnsignedCredential = await commonNetworkMember.buildRevocationListStatus(unsignedCredential)
 
     const affinityOptions = Object.assign({}, fullOptions, { apiKey: fullOptions.accessApiKey })
     const affinity = new Affinity(affinityOptions)
@@ -409,7 +390,7 @@ describe('CommonNetworkMember', () => {
     const sucessResult = await affinity.validateCredential(createdCredential)
     expect(sucessResult.result).to.equal(true)
 
-    await commonNetworkMember.revokeCredential(revokableUnsignedCredential.id, 'Status changed', accessToken)
+    await commonNetworkMember.revokeCredential(revokableUnsignedCredential.id, 'Status changed')
 
     const failResult = await affinity.validateCredential(createdCredential)
     expect(failResult.result).to.equal(false)